+++ packages/front-end/components/Features/SDKConnections/SDKLanguageLogo.tsx
@@ -389,3 +389,24 @@ export function getConnectionLanguageFilter(
   }
   return "all";
 }
+
+export function getPackageRepositoryName(url: string): string {
+  const repositoryMap: Record<string, string> = {
+    "npmjs.com": "NPM",
+    "pypi.org": "PyPI",
+    "rubygems.org": "RubyGems",
+    "packagist.org": "Packagist",
+    "jitpack.io": "JitPack",
+    "nuget.org": "NuGet",
+    "pkg.go.dev": "Go Modules",
+    "hex.pm": "Hex",
+    "swiftpackageindex.com": "Swift Package Index",
+    "mvnrepository.com": "Maven",
+    "pub.dev": "pub.dev",
+  };
+
+  for (const [domain, name] of Object.entries(repositoryMap)) {
+    if (url.includes(domain)) return name;
+  }
+  return "Package Repository";
+}
